+++ changelog.md
@@ -3,8 +3,13 @@ All notable changes to this project will be documented in this file.
 
 This project attempts to adhere to [Semantic Versioning](http://semver.org/).
 
-## [4.7.2] - [unreleased]
-### Changed
+## [4.8.0] - [unreleased]
+### Added
+- added `FugueCollectors` to complement fugue types with Stream API support which includes:
+  * `toEitherRight` collector (Similar to `Eithers#sequenceRight`)
+  * `toEitherLeft` collector (Similar to `Eithers#sequenceLeft`)
+  * `flatten` collector which flattens stream of `Option`s into stream of values filtering `None`s
+  * `toTrySuccess` collector (Similar to `Try#sequence`)
 
 ## [4.7.1] - [2018-07-19]
 ### Changed